+++ app/src/main/java/com/topjohnwu/magisk/ui/home/HomeFragment.kt
@@ -3,7 +3,6 @@ package com.topjohnwu.magisk.ui.home
 import android.view.Menu
 import android.view.MenuInflater
 import android.view.MenuItem
-import androidx.core.graphics.Insets
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentHomeMd2Binding
 import com.topjohnwu.magisk.model.navigation.Navigation
@@ -14,8 +13,6 @@ class HomeFragment : CompatFragment<HomeViewModel, FragmentHomeMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_home_md2
     override val viewModel by viewModel<HomeViewModel>()
-
-    override fun consumeSystemWindowInsets(insets: Insets) = insets
 
     override fun onStart() {
         super.onStart()
