+++ app/src/main/java/com/topjohnwu/magisk/ui/settings/SettingsFragment.kt
@@ -2,7 +2,6 @@ package com.topjohnwu.magisk.ui.settings
 
 import android.os.Bundle
 import android.view.View
-import androidx.core.graphics.Insets
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentSettingsMd2Binding
 import com.topjohnwu.magisk.ui.base.CompatFragment
@@ -13,8 +12,6 @@ class SettingsFragment : CompatFragment<SettingsViewModel, FragmentSettingsMd2Bi
 
     override val layoutRes = R.layout.fragment_settings_md2
     override val viewModel by viewModel<SettingsViewModel>()
-
-    override fun consumeSystemWindowInsets(insets: Insets) = insets
 
     override fun onStart() {
         super.onStart()
