+++ app/src/main/java/com/topjohnwu/magisk/ui/theme/ThemeFragment.kt
@@ -1,6 +1,5 @@
 package com.topjohnwu.magisk.ui.theme
 
-import androidx.core.graphics.Insets
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentThemeMd2Binding
 import com.topjohnwu.magisk.ui.base.CompatFragment
@@ -10,8 +9,6 @@ class ThemeFragment : CompatFragment<ThemeViewModel, FragmentThemeMd2Binding>() 
 
     override val layoutRes = R.layout.fragment_theme_md2
     override val viewModel by viewModel<ThemeViewModel>()
-
-    override fun consumeSystemWindowInsets(insets: Insets) = insets
 
     override fun onStart() {
         super.onStart()
