+++ app/src/main/res/layout/activity_flash.xml
@@ -16,20 +16,27 @@
         android:layout_height="match_parent">
 
         <com.google.android.material.appbar.AppBarLayout
+            style="@style/WidgetFoundation.Appbar"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:theme="@style/AppBarLayoutTheme.Flashing">
+            android:layout_height="wrap_content">
 
-            <androidx.appcompat.widget.Toolbar
+            <!--todo(diareuse) replace this with menus and navigation icon like a normal human being-->
+
+            <com.google.android.material.appbar.MaterialToolbar
+                style="@style/WidgetFoundation.Toolbar"
                 android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
+                android:layout_height="wrap_content"
+                android:minHeight="?attr/actionBarSize"
                 app:contentInsetLeft="0dp"
                 app:contentInsetStart="0dp"
-                app:popupTheme="@style/ToolbarPopupTheme.Flashing">
+                app:layout_scrollFlags="scroll|enterAlways">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
+                    android:layout_height="?attr/actionBarSize"
+                    android:layout_marginLeft="@{viewModel.insets.left}"
+                    android:layout_marginTop="@{viewModel.insets.top}"
+                    android:layout_marginRight="@{viewModel.insets.right}"
                     android:animateLayoutChanges="true">
 
                     <FrameLayout
@@ -46,7 +53,7 @@
                             android:layout_gravity="center"
                             android:background="@android:color/transparent"
                             app:srcCompat="@drawable/ic_back"
-                            app:tint="@android:color/white" />
+                            app:tint="?colorOnSurface" />
 
                     </FrameLayout>
 
@@ -57,7 +64,7 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:fontFamily="monospace"
-                        android:textColor="@android:color/white"
+                        android:textColor="?colorOnSurface"
                         app:layout_constraintBottom_toBottomOf="parent"
                         app:layout_constraintLeft_toLeftOf="parent"
                         app:layout_constraintRight_toRightOf="parent"
@@ -78,74 +85,46 @@
                             android:layout_gravity="center"
                             android:background="@android:color/transparent"
                             app:srcCompat="@drawable/ic_save_compat"
-                            app:tint="@android:color/white" />
+                            app:tint="?colorOnSurface" />
 
                     </FrameLayout>
 
                 </androidx.constraintlayout.widget.ConstraintLayout>
 
-            </androidx.appcompat.widget.Toolbar>
+            </com.google.android.material.appbar.MaterialToolbar>
 
         </com.google.android.material.appbar.AppBarLayout>
 
-        <androidx.constraintlayout.widget.ConstraintLayout
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/flash_content"
+            itemBinding="@{viewModel.itemBinding}"
+            items="@{viewModel.items}"
+            scrollToLast="@{true}"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior">
+            android:clipToPadding="false"
+            android:orientation="vertical"
+            android:paddingLeft="@{viewModel.insets.left}"
+            android:paddingRight="@{viewModel.insets.right}"
+            android:paddingBottom="@{viewModel.insets.bottom}"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            app:layout_behavior="@string/appbar_scrolling_view_behavior"
+            tools:listitem="@layout/item_console" />
 
-            <androidx.recyclerview.widget.RecyclerView
-                itemBinding="@{viewModel.itemBinding}"
-                items="@{viewModel.items}"
-                android:id="@+id/flash_content"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="vertical"
-                scrollToLast="@{true}"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                tools:listitem="@layout/item_console" />
-
-            <com.google.android.material.card.MaterialCardView
-                invisibleScale="@{!viewModel.loaded || !viewModel.canShowReboot}"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_margin="@dimen/margin_generic"
-                app:cardBackgroundColor="@color/colorSecondary"
-                app:cardCornerRadius="26dp"
-                app:cardPreventCornerOverlap="true"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintRight_toRightOf="parent">
-
-                <LinearLayout
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:animateLayoutChanges="true"
-                    android:background="?attr/selectableItemBackground"
-                    android:gravity="center"
-                    android:onClick="@{() -> viewModel.restartPressed()}"
-                    android:padding="@dimen/margin_generic_half">
-
-                    <androidx.appcompat.widget.AppCompatImageView
-                        style="@style/Widget.Icon"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:background="@android:color/transparent"
-                        app:srcCompat="@drawable/ic_restart"
-                        app:tint="@color/colorTextTinted" />
-
-                    <androidx.appcompat.widget.AppCompatTextView
-                        style="@style/Widget.Text.Emphasize.Tinted"
-                        gone="@{!viewModel.showRestartTitle}"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginLeft="@dimen/margin_generic_half"
-                        android:paddingRight="@dimen/margin_generic_half"
-                        android:text="@string/reboot" />
-
-                </LinearLayout>
-
-            </com.google.android.material.card.MaterialCardView>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
+        <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
+            invisibleScale="@{!viewModel.loaded || !viewModel.canShowReboot}"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom|end"
+            android:layout_margin="@dimen/l1"
+            android:layout_marginBottom="@{(int) @dimen/l1 + viewModel.insets.bottom}"
+            android:onClick="@{() -> viewModel.restartPressed()}"
+            android:text="@string/reboot"
+            android:textAllCaps="false"
+            android:textStyle="bold"
+            app:backgroundTint="?colorPrimary"
+            app:icon="@drawable/ic_restart"
+            app:tint="?colorOnPrimary" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
