+++ app/src/main/res/values-night/styles.xml
@@ -1,17 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="ThemeFoundation" parent="Theme.MaterialComponents.NoActionBar" />
-
-    <style name="MagiskTheme.SUFoundation" parent="Theme.MaterialComponents.Dialog" />
-
-    <style name="MagiskTheme" parent="ThemeFoundation.Colored.ExtraProps">
-        <!--All of these attributes are deprecated and will be removed-->
-        <item name="cardStyle">@style/CardViewStyle.Dark</item>
-        <item name="imageColorTint">@color/dark_secondary_text</item>
-        <item name="colorControl">?android:attr/textColorSecondary</item>
-        <item name="colorAccentFallback">@color/accentFallback</item>
-        <item name="android:windowBackground">@android:color/black</item>
-    </style>
-
 </resources>
