+++ app/src/main/res/values/styles.xml
@@ -1,65 +1,5 @@
 <resources>
 
-    <style name="ThemeFoundation" parent="Theme.MaterialComponents.Light.NoActionBar" />
-
-    <style name="ThemeFoundation.Colored">
-        <item name="colorPrimary">@color/colorPrimary</item>
-        <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
-        <item name="colorPrimaryVariant">@color/colorPrimaryDark</item>
-        <item name="colorAccent">@color/colorSecondary</item>
-        <item name="colorSecondary">@color/colorSecondary</item>
-        <item name="colorSecondaryVariant">@color/colorSecondaryDark</item>
-        <item name="android:windowBackground">@android:color/white</item>
-    </style>
-
-    <style name="ThemeFoundation.Colored.ExtraProps" />
-
-    <style name="MagiskTheme" parent="ThemeFoundation.Colored.ExtraProps">
-        <!--All of these attributes are deprecated and will be removed-->
-        <item name="cardStyle">@style/CardViewStyle.Light</item>
-        <item name="imageColorTint">@color/icon_grey</item>
-        <item name="colorControl">?android:attr/textColorSecondary</item>
-        <item name="colorAccentFallback">@color/accentFallback</item>
-    </style>
-
-    <style name="MagiskTheme.Flashing" parent="Theme.MaterialComponents.NoActionBar">
-        <item name="colorPrimary">@android:color/black</item>
-        <item name="colorPrimaryDark">@android:color/black</item>
-        <item name="colorAccent">@color/colorSecondary</item>
-        <item name="colorSecondary">@color/colorSecondary</item>
-        <item name="colorSecondaryVariant">@color/colorSecondaryDark</item>
-        <item name="android:windowBackground">@android:color/black</item>
-    </style>
-
-    <style name="MagiskTheme.SUFoundation" parent="Theme.MaterialComponents.Light.Dialog" />
-
-    <style name="MagiskTheme.SU" parent="MagiskTheme.SUFoundation">
-        <item name="android:windowIsFloating">true</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="windowActionBar">false</item>
-        <item name="windowNoTitle">true</item>
-    </style>
-
-    <!--Overlays-->
-
-    <style name="AppBarLayoutTheme" parent="ThemeOverlay.AppCompat.Dark" />
-
-    <style name="AppBarLayoutTheme.Flashing" parent="ThemeOverlay.AppCompat.Dark" />
-
-    <style name="ToolbarPopupTheme" parent="ThemeOverlay.AppCompat.Light" />
-
-    <style name="ToolbarPopupTheme.Flashing" parent="ThemeOverlay.AppCompat.Dark" />
-
-    <!--CardView-->
-
-    <style name="CardViewStyle.Dark" parent="CardView">
-        <item name="cardBackgroundColor">@color/card_background_color_dark</item>
-    </style>
-
-    <style name="CardViewStyle.Light" parent="CardView">
-        <item name="cardBackgroundColor">@color/card_background_color_light</item>
-    </style>
-
     <style name="SplashTheme" parent="SplashThemeBase"/>
 
 </resources>
